Moves testing assets in the instance-mutater API server package.
package_test.go is now package-local, instead of being in the *_test
package. This should be the pattern for all testing, facilitating the
current functionality of export_test files and white/black box testing.

export.go is deleted and its contents moved to package_test.go. Files
without a "_test" suffix are not excluded from normal builds, so we need
to avoid this.

A test has been added to ensure that charms with a nil profile do not
cause a panic. It covers the change added in #11304.
